% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{check_economies_names_codes}
\alias{check_economies_names_codes}
\title{Auxiliary function to check that a vector of country names contains valid country codes}
\usage{
check_economies_names_codes(.countries, .economies)
}
\arguments{
\item{.countries}{Vector of character strings containing the countries codes' to be checked.}

\item{.economies}{Vector of character strings containing the economies codes' to be checked.}
}
\description{
Auxiliary function to check that a vector of country names contains valid country codes
}
