% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{check_partner_economies}
\alias{check_partner_economies}
\title{Auxiliary function to check that a vector of country names contains valid country codes in the list of partner economies. Wraps `check_economies_names_codes()`}
\usage{
check_partner_economies(.countries)
}
\arguments{
\item{.countries}{Vector of character strings containing the countries codes' to be checked.}
}
\description{
Auxiliary function to check that a vector of country names contains valid country codes in the list of partner economies. Wraps `check_economies_names_codes()`
}
