% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{create_cache}
\alias{create_cache}
\title{Create a local `cachem` object to store the local cache, either in disk or memory.}
\usage{
create_cache(type = "disk")
}
\arguments{
\item{type}{Character string. Choose between types of cache to be created. Either 'disk' or 'memory'. Default is 'disk'}
}
\description{
Create a local `cachem` object to store the local cache, either in disk or memory.
}
