% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_api_key}
\alias{get_api_key}
\title{Retrieve the WTO API key from environment variable WTO_R_API_KEY}
\usage{
get_api_key()
}
\description{
Retrieve the WTO API key from environment variable WTO_R_API_KEY
}
