% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries_helper_functions.R
\name{get_bilateral_goods_trade}
\alias{get_bilateral_goods_trade}
\title{Helper function for `get_timeseries_data()`: retrieve bilateral goods trade between two reporting members}
\usage{
get_bilateral_goods_trade(
  .economy,
  .partner,
  .full_names = T,
  .last_period = T,
  .hs6_products = F,
  .products = "all",
  .nocache = F
)
}
\arguments{
\item{.economy}{Character string. Reporting economy code or name.}

\item{.partner}{Character string. Partner economy code or name.}

\item{.full_names}{Logical. Include a column called "full_name" containing the description for the HS6 codes.}

\item{.last_period}{Logical. Keep only values from the most recent period available. Default is TRUE.}

\item{.hs6_products}{Logical. Whether to return values by HS6 code Default is MTN products}

\item{.products}{Character string. Products to filter for. Default is "all", which results in no filtering.}

\item{.nocache}{Logical. TRUE to disable caching of results.}
}
\value{
A tibble containing the full list of NMF tariffs applied.
}
\description{
Helper function for `get_timeseries_data()`: retrieve bilateral goods trade between two reporting members
}
