% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_cached_object}
\alias{get_cached_object}
\title{Retrieve an object from local cache.}
\usage{
get_cached_object(key)
}
\arguments{
\item{key}{Character string. Cached object key.}
}
\description{
Retrieve an object from local cache.
}
