% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_frequencies.R
\name{get_frequencies}
\alias{get_frequencies}
\title{Retrieve the available data frequencies.}
\usage{
get_frequencies(lang = "1", nocache = F)
}
\arguments{
\item{lang}{Integer. Set to 1 for English, 2 for French or 3 for Spanish}

\item{nocache}{Logical. If TRUE, disables retrieval of results from local cache.}
}
\description{
Retrieve the available data frequencies.
}
\examples{
get_frequencies()
}
