% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geographical_regions.R
\name{get_geographical_regions}
\alias{get_geographical_regions}
\title{Retrieve the list of available geographical regions to be passed as reporting or partner economies parameters.}
\usage{
get_geographical_regions(lang = "1", nocache = F)
}
\arguments{
\item{lang}{Integer. Set to 1 for English, 2 for French or 3 for Spanish}

\item{nocache}{Logical. If TRUE, disables retrieval of results from local cache.}
}
\value{
A tibble containing the list of geographical regions and their corresponding codes.
}
\description{
Retrieve the list of available geographical regions to be passed as reporting or partner economies parameters.
}
\examples{
get_geographical_regions()
}
