% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries_helper_functions.R
\name{get_group_members}
\alias{get_group_members}
\title{Get a list of the economies belonging to a given group}
\usage{
get_group_members(
  .type = "reporting",
  .group = NULL,
  .lang = "1",
  .nocache = F
)
}
\arguments{
\item{.type}{Character string. Either "reporting" or "partner". Default is "reporting".}

\item{.group}{Character string. Code of the group whose members are to be retrieved. Default is NULL. Retrieve a list of codes with `get_groups()`}

\item{.lang}{Character string. Language of the returned output. Default is "1". "1" for English, "2" for French and "3" for Spanish.}

\item{.nocache}{Logical. TRUE to disable caching of results. Default is FALSE.}
}
\description{
Get a list of the economies belonging to a given group
}
