% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groups.R
\name{get_groups}
\alias{get_groups}
\title{Retrieve the list of available groups of economies to be passed as reporting or partner economies parameters.}
\usage{
get_groups(lang = "1", nocache = F)
}
\arguments{
\item{lang}{Numeric. Set to 1 for English, 2 for French or 3 for Spanish}

\item{nocache}{Logical. If TRUE, disables retrieval of results from local cache.}
}
\value{
A tibble containing the list of available groups of economies and their corresponding codes.
}
\description{
Retrieve the list of available groups of economies to be passed as reporting or partner economies parameters.
}
\examples{
get_groups()
}
