% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicator_categories.R
\name{get_indicator_categories}
\alias{get_indicator_categories}
\title{Retrieve a list of categories of available Timeseries API indicators.}
\usage{
get_indicator_categories(lang = "1", nocache = F)
}
\arguments{
\item{lang}{Numeric. Set to 1 for English, 2 for French or 3 for Spanish}

\item{nocache}{Logical. If TRUE, disables retrieval of results from local cache.}
}
\value{
A tibble containing the codes of indicator categories and their description.
}
\description{
Retrieve a list of categories of available Timeseries API indicators.
}
\examples{
get_indicator_categories()
}
