% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_product_classifications.R
\name{get_product_classifications}
\alias{get_product_classifications}
\title{Retrieve a list of valid product classifications codes and their descriptions.}
\usage{
get_product_classifications(lang = "1", nocache = F)
}
\arguments{
\item{lang}{Integer. Set to 1 for English, 2 for French or 3 for Spanish}

\item{nocache}{Logical. If TRUE, disables retrieval of results from local cache.}
}
\value{
A tibble containing the list of valid product classifications.
}
\description{
Retrieve a list of valid product classifications codes and their descriptions.
}
\examples{
get_product_classifications()
}
