% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_products_sectors.R
\name{get_products_sectors}
\alias{get_products_sectors}
\title{Retrieve a list of valid codes for products (goods) and sectors (services).}
\usage{
get_products_sectors(pc = "all", lang = "1", nocache = F)
}
\arguments{
\item{pc}{Character string. Product classification to filter for. See `get_product_classification()`. Default is 'all'.}

\item{lang}{Numeric. Set to 1 for English, 2 for French or 3 for Spanish}

\item{nocache}{Logical. If TRUE, disables retrieval of results from local cache.}
}
\value{
A tibble containing the available periods.
}
\description{
Retrieve a list of valid codes for products (goods) and sectors (services).
}
\examples{
get_periods()
}
