% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qr_hs_versions.R
\name{get_qr_hs_versions}
\alias{get_qr_hs_versions}
\title{Retrieve the list of available HS versions for the Quantitative Restrictions database.}
\usage{
get_qr_hs_versions(nocache = F)
}
\arguments{
\item{nocache}{Logical. If TRUE, disables retrieval of results from local cache.}
}
\value{
A tibble containing the list of valid codes for available HS version.
}
\description{
Retrieve the list of available HS versions for the Quantitative Restrictions database.
}
\examples{
get_qr_hs_versions()
}
