% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qr_id.R
\name{get_qr_id}
\alias{get_qr_id}
\title{Retrieve quantitative restrictions by id}
\usage{
get_qr_id(qr_id, nocache = F)
}
\arguments{
\item{qr_id}{Id of the QR}

\item{nocache}{Logical. If TRUE, disables retrieval of results from local cache.}
}
\value{
A tibble containing the details of the quantitative restriction whose id matches the argument.
}
\description{
Retrieve quantitative restrictions by id
}
\examples{
get_qr_id(qr_id="")
}
