% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qr_members.R
\name{get_qr_members}
\alias{get_qr_members}
\title{Retrieve a list of member countries names and codes that can be used as arguments to other QR functions.}
\usage{
get_qr_members(member_code = NULL, name = NULL, page = NULL, nocache = F)
}
\arguments{
\item{member_code}{Code of the member country.}

\item{name}{Name of the member country.}

\item{page}{Page to be retrieved}

\item{nocache}{Logical. If TRUE, disables retrieval of results from local cache.}
}
\value{
A tibble containing the name and code information for the requested country.
}
\description{
Retrieve a list of member countries names and codes that can be used as arguments to other QR functions.
}
\examples{
get_qr_members(name="Afghanistan")
}
