% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qr_products.R
\name{get_qr_products}
\alias{get_qr_products}
\title{Retrieve product lists from the Quantitative Restrictions database}
\usage{
get_qr_products(
  hs_version,
  code = NULL,
  page = NULL,
  description = NULL,
  nocache = F
)
}
\arguments{
\item{hs_version}{Character string. One of the following strings: "h1", "h2", "h3", "h4", "h5", "h6"}

\item{code}{Character string. The product code to fully or partially match}

\item{page}{Integer. Page number}

\item{description}{Character string. String to be searched for in the product description}

\item{nocache}{Logical. If TRUE, disables retrieval of results from local cache.}
}
\description{
Retrieve product lists from the Quantitative Restrictions database
}
\examples{
get_qr_products(hs_version="h1", code="65")
}
