% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reporting_economies.R
\name{get_reporting_economies}
\alias{get_reporting_economies}
\title{Retrieve the list of reporting economies and their codes.}
\usage{
get_reporting_economies(
  ig = "all",
  reg = "all",
  gp = "all",
  lang = "1",
  nocache = F
)
}
\arguments{
\item{ig}{Character string. Either 'all', 'group' (for groups of economies) or 'individual' (for individual economies).}

\item{reg}{Character string. Region to filter for. See `get_geographical_regions()`. Default is 'all'.}

\item{gp}{Character string. Group of economies to filter for. See `get_groups()`. Default is 'all'.}

\item{lang}{Integer. Set to 1 for English, 2 for French or 3 for Spanish}

\item{nocache}{Logical. If TRUE, disables retrieval of results from local cache.}
}
\value{
A tibble containing the list of reporting economies and their corresponding country codes.
}
\description{
Retrieve the list of reporting economies and their codes.
}
\examples{
get_partner_economies()
}
