% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries_helper_functions.R
\name{get_services_imports}
\alias{get_services_imports}
\title{Helper function for `get_timeseries_data()`: retrieve services imports of a given economy}
\usage{
get_services_imports(
  .economy,
  .partner = NULL,
  .full_names = T,
  .last_period = T,
  .nocache = F
)
}
\arguments{
\item{.economy}{Character string. Reporting economy code or name.}

\item{.partner}{Character string. Partner economy code or name.}

\item{.full_names}{Logical. Include a column called "full_name" containing the description for the HS6 codes.}

\item{.last_period}{Logical. Keep only values from the most recent period available. Default is TRUE.}

\item{.nocache}{Logical. TRUE to disable caching of results.}
}
\value{
A tibble containing the full list of MFN tariffs applied.
}
\description{
Helper function for `get_timeseries_data()`: retrieve services imports of a given economy
}
