% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries_helper_functions.R
\name{get_tariffs_nmf}
\alias{get_tariffs_nmf}
\title{Helper function for `get_timeseries_data()`: retrieve all NMF tariffs for a given reporting economy.}
\usage{
get_tariffs_nmf(.economy, .full_names = T)
}
\arguments{
\item{.economy}{Character string. Reporting economy code or name.}

\item{.full_names}{Logical. Include a column called "full_name" containing the description for the HS6 codes.}
}
\value{
A tibble containing the full list of NMF tariffs applied.
}
\description{
Helper function for `get_timeseries_data()`: retrieve all NMF tariffs for a given reporting economy.
}
