% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries_available_indicators.R
\name{get_timeseries_available_indicators}
\alias{get_timeseries_available_indicators}
\title{Retrieve the list of available indicators for the WTO Timeseries API}
\usage{
get_timeseries_available_indicators(
  i = "all",
  name = "",
  t = "all",
  pc = "all",
  tp = "all",
  frq = "all",
  lang = "1",
  nocache = F
)
}
\arguments{
\item{i}{Character string. Either 'all' or a string to filter indicators' description for.}

\item{name}{Character string. Indicator name or part of it.}

\item{t}{Character string. Topics to filter for, separated by commas.}

\item{pc}{Character string. Product classifications: must be one of the following: 'all', 'none', or a list of comma separated codes.}

\item{tp}{Character string. Filter for indicators with a trade partner paramenter. Either 'all', 'yes' or no.}

\item{frq}{Character string. Filter for the frequency available. Either 'all' or 'A', 'S', 'Q', 'M' (anually, semi-anually, quarterly or monthly).}

\item{lang}{Integer. Set to 1 for English, 2 for French or 3 for Spanish}

\item{nocache}{Logical. If TRUE, disables retrieval of results from local cache.}
}
\value{
A tibble containing the available indicators with the specified filters.
}
\description{
Retrieve the list of available indicators for the WTO Timeseries API
}
\examples{
get_timeseries_available_indicators()
}
