% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries_metadata.R
\name{get_timeseries_metadata}
\alias{get_timeseries_metadata}
\title{Retrieve metadata for the Timeseries API indicators.}
\usage{
get_timeseries_metadata(
  code,
  reporting_economies = "all",
  partner_economies = "all",
  product_classification_code = "all",
  subproducts_subsectors = TRUE,
  lang = 1,
  nocache = F
)
}
\arguments{
\item{code}{indicator code. Required.}

\item{reporting_economies}{Character string. A vector or a scalar containing the codes and/or names of the reporting economies.}

\item{partner_economies}{Character string. A vector or a scalar containing the codes and/or names of the partner economies. Not all indicators allow for this parameter.}

\item{product_classification_code}{Character string. Either TRUE or FALSE depending on whether to include or not subproducts and subsectors.}

\item{subproducts_subsectors}{Logical. Either TRUE or FALSE depending on whether to include or not subproducts and subsectors.}

\item{lang}{Integer. Set to 1 for English, 2 for French or 3 for Spanish}

\item{nocache}{Logical. If TRUE, disables retrieval of results from local cache.}
}
\description{
Retrieve metadata for the Timeseries API indicators.
}
