% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topics.R
\name{get_topics}
\alias{get_topics}
\title{Retrieve a list of topics grouping indicators.}
\usage{
get_topics(lang = "1", nocache = F)
}
\arguments{
\item{lang}{Numeric. Set to 1 for English, 2 for French or 3 for Spanish}

\item{nocache}{Logical. If TRUE, disables retrieval of results from local cache.}
}
\description{
Retrieve a list of topics grouping indicators.
}
