% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_units.R
\name{get_units}
\alias{get_units}
\title{Retrieve a list of valid for the units in which the data may be retrieved.}
\usage{
get_units(lang = "1", nocache = F)
}
\arguments{
\item{lang}{Numeric. Set to 1 for English, 2 for French or 3 for Spanish}

\item{nocache}{Logical. If TRUE, disables retrieval of results from local cache.}
}
\description{
Retrieve a list of valid for the units in which the data may be retrieved.
}
