% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_value_flags.R
\name{get_value_flags}
\alias{get_value_flags}
\title{Retrive a list of valid codes for value flags. Value flags are predefined metadata for values, such as estimation, ad valorem, forecasted values...}
\usage{
get_value_flags(lang = "1", pc = "all", nocache = F)
}
\arguments{
\item{lang}{Numeric. Set to 1 for English, 2 for French or 3 for Spanish}

\item{pc}{Character string. Product classifications: must be one of the following: 'all', 'none', or a list of comma separated codes.}

\item{nocache}{Logical. If TRUE, disables retrieval of results from local cache.}
}
\description{
Retrive a list of valid codes for value flags. Value flags are predefined metadata for values, such as estimation, ad valorem, forecasted values...
}
