% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries_helper_functions.R
\name{get_wto_ts}
\alias{get_wto_ts}
\title{Generic helper function for `get_timeseries_data()`}
\usage{
get_wto_ts(
  .code,
  .economy,
  .partner = NULL,
  .full_names = T,
  .last_period = T,
  .year = NULL,
  .products_or_sectors = "all",
  .nocache = F,
  .debug_info = F
)
}
\arguments{
\item{.code}{Character string. The timeseries code to be retrieved.}

\item{.economy}{Character string. Reporting economy ISO three digit or three letters code, or its name in `get_reporting_economies()`.}

\item{.partner}{Character string. Partner economy ISO three digit or three letters code, or its name in `get_partner_economies()`.}

\item{.full_names}{Logical. Include a column called "full_name" containing the description for the HS6 codes. Default is FALSE.}

\item{.last_period}{Logical. Keep only values from the most recent period available. Default is TRUE.}

\item{.year}{Integer. Select year to download. If `NULL` it retrieves all the periods available.}

\item{.products_or_sectors}{Character string. Products or sectors to download. By default `"all"`.}

\item{.nocache}{Logical. TRUE to disable caching of results.}

\item{.debug_info}{Logical. Print debugging information. Default is FALSE.}
}
\description{
Provides a simple wrapper to `get_timeseries_data()`.
}
