% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_years.R
\name{get_years}
\alias{get_years}
\title{Retrieve a list of available years for the WTO Timeseries API}
\usage{
get_years(lang = "1", nocache = F)
}
\arguments{
\item{lang}{1 for English, 2 for French or 3 for Spanish}

\item{nocache}{Logical. If TRUE, disables retrieval of results from local cache.}
}
\description{
Retrieve a list of available years for the WTO Timeseries API
}
\examples{
get_years()
}
