% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries_helper_functions.R
\name{is_wto_member}
\alias{is_wto_member}
\title{Check whether a given economy is a member of the WTO}
\usage{
is_wto_member(.code, .nocache = "F")
}
\arguments{
\item{.code}{Character string. ISO code of the economy whose WTO member status is to be checked.}

\item{.nocache}{Logical. TRUE to disable caching of results. Default is FALSE.}
}
\value{
TRUE is the economy is a member of the WTO
}
\description{
Check whether a given economy is a member of the WTO
}
