% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{set_cached_object}
\alias{set_cached_object}
\title{Store an object in local cache.}
\usage{
set_cached_object(key, value)
}
\arguments{
\item{key}{Character string. Cached object key.}

\item{value}{Tibble. Cached object.}
}
\description{
Store an object in local cache.
}
