% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtor-package.R
\docType{package}
\name{wtor-package}
\alias{wtor}
\alias{wtor-package}
\title{wtor: A Client for the WTO API Written in R}
\description{
A client for the excellent World Trade Organization REST API. Implements functions to query most of the available endpoints.
}
\keyword{internal}
