% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{wtor_env}
\alias{wtor_env}
\title{Create an environment object in which to store the local cache object.}
\format{
An object of class \code{environment} of length 1.
}
\usage{
wtor_env
}
\description{
Create an environment object in which to store the local cache object.
}
\keyword{datasets}
