% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_periods.R
\name{get_periods}
\alias{get_periods}
\title{Retrieve available periods for the WTO Timeseries API}
\usage{
get_periods(lang = "1", nocache = F)
}
\arguments{
\item{lang}{Numeric. Set to 1 for English, 2 for French or 3 for Spanish}

\item{nocache}{Logical. If TRUE, disables retrieval of results from local cache.}
}
\value{
A tibble containing the available periods.
}
\description{
Retrieve available periods for the WTO Timeseries API
}
\examples{
get_periods()
}
