% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{get_timeseries_data}
\alias{get_timeseries_data}
\title{Retrieve timeseries data.}
\usage{
get_timeseries_data(
  code,
  reporting_economies,
  partner_economies = NULL,
  time_period = "default",
  products_or_sectors = "default",
  subproducts_subsectors = FALSE,
  format_output = "csv",
  mode_output = "codes",
  decimals = "default",
  heading_style = "M",
  offset = 0,
  max_records = NULL,
  meta = FALSE,
  lang = "1",
  noitemcount = FALSE,
  nocache = F,
  nopagination = F,
  pageitems = 999999,
  request_max_attempts = 10,
  debug_info = FALSE
)
}
\arguments{
\item{code}{Character string. indicator code. Required.}

\item{reporting_economies}{Character string. A vector or a scalar containing the codes and/or names of the reporting economies.}

\item{partner_economies}{Character string. A vector or a scalar containing the codes and/or names of the partner economies. Not all indicators allow for this parameter.}

\item{time_period}{Character string. A string containing either "default", "all", or specific periods according to the format described in https://apiportal.wto.org/api-details#api=version1&operation=post-data}

\item{products_or_sectors}{Character string. A string containing either "default", "all", a specific product classification such as HS2, HS4, HS6, or a comma separated list of product codes belonging to AG,AGFOFI,MAIS,...}

\item{subproducts_subsectors}{Logical. Either TRUE or FALSE depending on whether to include or not subproducts and subsectors.}

\item{format_output}{Character string. jEither "csv" or "json", depending on the output format in which to obtain the response to the POST request. It does not have any impact on the function returned dataframe.}

\item{mode_output}{Character string. Either "codes" (by default) or "full", depending on whether the columns in the returned dataframe will contain}

\item{decimals}{Character string. Either "default" or a string containing the number of decimals that the output should contain.}

\item{heading_style}{Character string. Either "H" for human-readable headers and "M" for machine-readable codes.}

\item{offset}{Character string. Number of datapoints to offset in the request. Usefull if manual pagination is to be implemented.}

\item{max_records}{Character string. Maximum number of rows to return. Default is NULL. If NULL, the maximum number of rows will equal the number of datapoints returned by get_timeseries_data_count()}

\item{meta}{Logical. TRUE to include metadata.}

\item{lang}{Integer. Set to 1 for English, 2 for French or 3 for Spanish}

\item{noitemcount}{Logical. Set to TRUE to disabling querying for the number of data points before data retrieval. Default is FALSE. Setting this argument to TRUE may interfere with pagination.}

\item{nocache}{Logical. TRUE to disable caching of results.}

\item{nopagination}{Logical. TRUE to disable pagination of requests.}

\item{pageitems}{Numeric. Number of rows per paginated request. By default 10.000}

\item{request_max_attempts}{Numeric. Maximum number of request attempts.}

\item{debug_info}{Logical. Print debugging information. Default is FALSE.}
}
\value{
A tibble containing the request data.
}
\description{
Retrieve timeseries data.
}
